Shared the duplicated Behat scope factories through 'BehatScopeTrait'.
