+++ tests/phpunit/Unit/ScreenshotContextAnimationTest.php
@@ -4,15 +4,11 @@ declare(strict_types=1);
 
 namespace DrevOps\BehatScreenshot\Tests\Unit;
 
-use Behat\Behat\Hook\Scope\AfterScenarioScope;
-use Behat\Behat\Hook\Scope\AfterStepScope;
 use Behat\Behat\Hook\Scope\BeforeScenarioScope;
-use Behat\Behat\Tester\Result\StepResult;
 use Behat\Gherkin\Node\FeatureNode;
 use Behat\Gherkin\Node\ScenarioInterface;
-use Behat\Gherkin\Node\StepNode;
 use Behat\Testwork\Environment\Environment;
-use Behat\Testwork\Tester\Result\TestResult;
+use DrevOps\BehatScreenshot\Tests\Traits\BehatScopeTrait;
 use DrevOps\BehatScreenshot\Tests\Traits\ReflectionTrait;
 use DrevOps\BehatScreenshotExtension\AnimatedGif;
 use DrevOps\BehatScreenshotExtension\Context\ScreenshotContext;
@@ -26,6 +22,7 @@ use PHPUnit\Framework\TestCase;
 #[CoversClass(ScreenshotContext::class)]
 class ScreenshotContextAnimationTest extends TestCase {
 
+  use BehatScopeTrait;
   use ReflectionTrait;
 
   #[DataProvider('dataProviderBeforeScenarioCheckScreenshotsTagSetsFlagsFromTagsAndConfig')]
@@ -279,42 +276,6 @@ class ScreenshotContextAnimationTest extends TestCase {
       'non-numeric setting' => [['frame_delay' => 'fast'], 'frame_delay', 500, 500],
       'null setting' => [['frame_delay' => NULL], 'frame_delay', 500, 500],
     ];
-  }
-
-  /**
-   * Create an after step scope with the given result state.
-   *
-   * @param bool $passed
-   *   Whether the step passed.
-   *
-   * @return \Behat\Behat\Hook\Scope\AfterStepScope
-   *   After step scope.
-   */
-  protected function createAfterStepScope(bool $passed): AfterStepScope {
-    $result = $this->createMock(StepResult::class);
-    $result->method('isPassed')->willReturn($passed);
-
-    return new AfterStepScope($this->createMock(Environment::class), $this->createMock(FeatureNode::class), $this->createMock(StepNode::class), $result);
-  }
-
-  /**
-   * Create an after scenario scope.
-   *
-   * @param string|null $feature_file
-   *   Feature file path.
-   * @param int $scenario_line
-   *   Scenario line number.
-   *
-   * @return \Behat\Behat\Hook\Scope\AfterScenarioScope
-   *   After scenario scope.
-   */
-  protected function createAfterScenarioScope(?string $feature_file = NULL, int $scenario_line = 0): AfterScenarioScope {
-    $feature_node = $this->createMock(FeatureNode::class);
-    $feature_node->method('getFile')->willReturn($feature_file);
-    $scenario = $this->createMock(ScenarioInterface::class);
-    $scenario->method('getLine')->willReturn($scenario_line);
-
-    return new AfterScenarioScope($this->createMock(Environment::class), $feature_node, $scenario, $this->createMock(TestResult::class));
   }
 
 }
